* Update includes to be case sensitive * Rename Nanoshell.hpp -> NanoShell.hpp
