+++ source/EternalHeap.cpp
@@ -7,7 +7,7 @@
 //  Programmer(s):  iProgramInCpp (iprogramincpp@gmail.com)
 //  ***************************************************************
 
-#include <Nanoshell.hpp>
+#include <NanoShell.hpp>
 #include <EternalHeap.hpp>
 #include <Spinlock.hpp>
