+++ source/System.cpp
@@ -6,7 +6,7 @@
 //  ***************************************************************
 //  Programmer(s):  iProgramInCpp (iprogramincpp@gmail.com)
 //  ***************************************************************
-#include <Nanoshell.hpp>
+#include <NanoShell.hpp>
 
 extern "C" void AssertUnreachable(const char* src_file, int src_line)
 {
